+++ tests/test_title_main.py
@@ -6,7 +6,7 @@ from decimal import Decimal
 from cinemagoerng import model, web
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",), ("taglines",)])
+@pytest.mark.parametrize(("page",), [("reference",), ("taglines",)])
 @pytest.mark.parametrize(
     ("imdb_id",),
     [
@@ -18,7 +18,7 @@ def test_title_parser_should_set_imdb_id(page, imdb_id):
     assert parsed.imdb_id == imdb_id
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",), ("taglines",)])
+@pytest.mark.parametrize(("page",), [("reference",), ("taglines",)])
 @pytest.mark.parametrize(
     ("imdb_id", "type_"),
     [
@@ -40,7 +40,7 @@ def test_title_parser_should_instantiate_correct_type(page, imdb_id, type_):
     assert isinstance(parsed, type_)
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",), ("taglines",)])
+@pytest.mark.parametrize(("page",), [("reference",), ("taglines",)])
 @pytest.mark.parametrize(
     ("imdb_id", "title"),
     [
@@ -63,7 +63,7 @@ def test_title_parser_should_set_title_from_original_title(page, imdb_id, title)
     assert parsed.title == title
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",), ("taglines",)])
+@pytest.mark.parametrize(("page",), [("reference",), ("taglines",)])
 @pytest.mark.parametrize(
     ("imdb_id", "primary_image"),
     [
@@ -79,7 +79,7 @@ def test_title_parser_should_set_primary_image(page, imdb_id, primary_image):
     assert parsed.primary_image == primary_image
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",), ("taglines",)])
+@pytest.mark.parametrize(("page",), [("reference",), ("taglines",)])
 @pytest.mark.parametrize(
     ("imdb_id", "year"),
     [
@@ -95,7 +95,7 @@ def test_title_parser_should_set_year(page, imdb_id, year):
     assert parsed.year == year
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",), ("taglines",)])
+@pytest.mark.parametrize(("page",), [("reference",), ("taglines",)])
 @pytest.mark.parametrize(
     ("imdb_id", "end_year"),
     [
@@ -109,7 +109,7 @@ def test_title_parser_should_set_end_year(page, imdb_id, end_year):
     assert parsed.end_year == end_year
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",), ("taglines",)])
+@pytest.mark.parametrize(("page",), [("reference",), ("taglines",)])
 @pytest.mark.parametrize(
     ("imdb_id",),
     [
@@ -129,7 +129,7 @@ def test_title_parser_should_not_set_end_year_for_other_than_series(page, imdb_i
     assert not hasattr(parsed, "end_year")
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "country_codes"),
     [
@@ -142,7 +142,7 @@ def test_title_parser_should_set_country_codes(page, imdb_id, country_codes):
     assert parsed.country_codes == country_codes
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "language_codes"),
     [
@@ -156,7 +156,7 @@ def test_title_parser_should_set_language_codes(page, imdb_id, language_codes):
     assert parsed.language_codes == language_codes
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "runtime"),
     [
@@ -174,7 +174,7 @@ def test_title_parser_should_set_runtime(page, imdb_id, runtime):
     assert parsed.runtime == runtime
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id",),
     [
@@ -186,7 +186,7 @@ def test_title_parser_should_not_set_runtime_for_video_games(page, imdb_id):
     assert not hasattr(parsed, "runtime")
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "genres"),
     [
@@ -226,7 +226,7 @@ def test_title_parser_should_set_genres(page, imdb_id, genres):
     assert parsed.genres == genres
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "plot", "lang"),
     [
@@ -240,7 +240,7 @@ def test_title_parser_should_set_plot(page, imdb_id, plot, lang):
     assert parsed.plot[lang].startswith(plot)
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "rating"),
     [
@@ -253,7 +253,7 @@ def test_title_parser_should_set_rating(page, imdb_id, rating):
     assert (abs(parsed.rating - rating) < Decimal("0.3")) if rating is not None else (parsed.rating is None)
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "votes"),
     [
@@ -266,7 +266,7 @@ def test_title_parser_should_set_vote_count(page, imdb_id, votes):
     assert (parsed.vote_count >= votes) if votes > 0 else (parsed.vote_count == 0)
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "rank"),
     [
@@ -279,157 +279,7 @@ def test_title_parser_should_set_top_ranking(page, imdb_id, rank):
     assert (abs(parsed.top_ranking - rank) < 10) if rank is not None else (parsed.top_ranking is None)
 
 
-@pytest.mark.parametrize(
-    ("imdb_id", "n", "cast"),
-    [
-        (
-            "tt7045440",
-            1,
-            [  # David Bowie: Ziggy Stardust
-                ("nm0000309", "David Bowie", ["Self"], []),
-            ],
-        ),
-        (
-            "tt0101597",
-            2,
-            [  # Closet Land
-                ("nm0000656", "Madeleine Stowe", ["Victim"], []),
-                ("nm0000614", "Alan Rickman", ["Interrogator"], []),
-            ],
-        ),
-        (
-            "tt1000252",
-            12,
-            [  # Blink
-                ("nm0855039", "David Tennant", ["The Doctor"], []),
-                ("nm1303956", "Freema Agyeman", ["Martha Jones"], []),
-                ("nm1659547", "Carey Mulligan", ["Sally Sparrow"], []),
-                ("nm1164725", "Lucy Gaskell", ["Kathy Nightingale"], []),
-                ("nm1015511", "Finlay Robertson", ["Larry Nightingale"], []),
-                ("nm0134458", "Richard Cant", ["Malcolm Wainwright"], []),
-                ("nm0643394", "Michael Obiora", ["Billy Shipton"], []),
-                ("nm0537158", "Louis Mahoney", ["Old Billy"], []),
-                ("nm1631281", "Thomas Nelstrop", ["Ben Wainwright"], []),
-                ("nm2286323", "Ian Boldsworth", ["Banto"], []),
-                ("nm0768205", "Raymond Sawyer", ["Desk Sergeant"], ["as Ray Sawyer"]),
-                ("nm4495179", "Elen Thomas", ["Weeping Angel"], ["uncredited"]),
-            ],
-        ),
-        ("tt0133093", 18, []),  # The Matrix
-        ("tt3629794", 0, []),  # Aslan
-    ],
-)
-def test_title_main_parser_should_set_main_cast(imdb_id, n, cast):
-    parsed = web.get_title(imdb_id=imdb_id, page="main")
-    assert len(parsed.cast) == n
-    if len(cast) > 0:
-        assert [(credit.imdb_id, credit.name, credit.characters, credit.notes) for credit in parsed.cast] == cast
-
-
-@pytest.mark.parametrize(
-    ("imdb_id", "n", "directors"),
-    [
-        (
-            "tt1000252",
-            1,
-            [  # Blink
-                ("nm0531751", "Hettie Macdonald", None, []),
-            ],
-        ),
-        (
-            "tt0133093",
-            2,
-            [  # The Matrix
-                ("nm0905154", "Lana Wachowski", None, []),
-                ("nm0905152", "Lilly Wachowski", None, []),
-            ],
-        ),
-        (
-            "tt0092580",
-            3,
-            [  # Aria
-                ("nm0000265", "Robert Altman", None, []),
-                ("nm0000915", "Bruce Beresford", None, []),
-                ("nm0117317", "Bill Bryden", None, []),
-            ],
-        ),
-        ("tt3629794", 0, []),  # Aslan
-    ],
-)
-def test_title_main_parser_should_set_main_directors(imdb_id, n, directors):
-    parsed = web.get_title(imdb_id=imdb_id, page="main")
-    assert len(parsed.directors) == n
-    if len(directors) > 0:
-        assert [(credit.imdb_id, credit.name, credit.job, credit.notes) for credit in parsed.directors] == directors
-
-
-@pytest.mark.parametrize(
-    ("imdb_id", "n", "writers"),
-    [
-        (
-            "tt7045440",
-            1,
-            [  # David Bowie: Ziggy Stardust
-                ("nm0000309", "David Bowie", None, [])
-            ],
-        ),
-        (
-            "tt0133093",
-            2,
-            [  # The Matrix
-                ("nm0905152", "Lilly Wachowski", None, []),
-                ("nm0905154", "Lana Wachowski", None, []),
-            ],
-        ),
-        (
-            "tt0076786",
-            3,
-            [  # Suspiria
-                ("nm0000783", "Dario Argento", None, []),
-                ("nm0630453", "Daria Nicolodi", None, []),
-                ("nm0211063", "Thomas De Quincey", None, []),
-            ],
-        ),
-        ("tt0092580", 3, []),  # Aria
-        ("tt0365467", 0, []),  # Making 'The Matrix'
-    ],
-)
-def test_title_main_parser_should_set_main_writers(imdb_id, n, writers):
-    parsed = web.get_title(imdb_id=imdb_id, page="main")
-    assert len(parsed.writers) == n
-    if len(writers) > 0:
-        assert [(credit.imdb_id, credit.name, credit.job, credit.notes) for credit in parsed.writers] == writers
-
-
-@pytest.mark.parametrize(
-    ("imdb_id", "n", "creators"),
-    [
-        (
-            "tt0436992",
-            1,
-            [  # Doctor Who
-                ("nm0628285", "Sydney Newman", None, [])
-            ],
-        ),
-        (
-            "tt0445114",
-            2,
-            [  # Extras
-                ("nm0315041", "Ricky Gervais", None, []),
-                ("nm0580351", "Stephen Merchant", None, []),
-            ],
-        ),
-        ("tt0185906", 0, []),  # Band of Brothers (Mini-Series)
-    ],
-)
-def test_title_main_parser_should_set_main_creators_for_series(imdb_id, n, creators):
-    parsed = web.get_title(imdb_id=imdb_id, page="main")
-    assert len(parsed.creators) == n
-    if len(creators) > 0:
-        assert [(credit.imdb_id, credit.name, credit.job, credit.notes) for credit in parsed.creators] == creators
-
-
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "type_", "series_data"),
     [
@@ -443,7 +293,7 @@ def test_title_parser_should_set_series_for_episode(page, imdb_id, type_, series
     assert (parsed.series.imdb_id, parsed.series.title) == series_data
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "year", "end_year"),
     [
@@ -456,7 +306,7 @@ def test_title_parser_should_set_series_years_for_episode(page, imdb_id, year, e
     assert (parsed.series.year, parsed.series.end_year) == (year, end_year)
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "season", "episode"),
     [
@@ -468,7 +318,7 @@ def test_title_parser_should_set_season_and_episode_number_for_episode(page, imd
     assert (parsed.season, parsed.episode) == (season, episode)
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "release_date"),
     [
@@ -480,7 +330,7 @@ def test_title_parser_should_set_release_date_for_episode(page, imdb_id, release
     assert parsed.release_date == release_date
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "episode_id"),
     [
@@ -493,7 +343,7 @@ def test_title_parser_should_set_previous_episode_for_episode(page, imdb_id, epi
     assert parsed.previous_episode == episode_id
 
 
-@pytest.mark.parametrize(("page",), [("main",), ("reference",)])
+@pytest.mark.parametrize(("page",), [("reference",)])
 @pytest.mark.parametrize(
     ("imdb_id", "episode_id"),
     [